requirements.txt
Django==4.2.0
psycopg2-binary==2.9.6
python-decouple==3.8
Pillow==10.0.0
djangorestframework==3.14.0
django-cors-headers==4.2.0
supabase==1.0.3
reportlab==4.0.4
